+++ camerax/lib/src/widgets.dart
@@ -1,0 +1,1 @@
+export 'widgets/preview_widget.dart';