+++ camerax_ios/lib/src/view/preview_view_impl.dart
@@ -1,3 +1,5 @@
+import 'dart:ui';
+
 import 'package:camerax_ios/src/api.dart';
 import 'package:camerax_ios/src/common.dart';
 import 'package:camerax_ios/src/view.dart';
@@ -14,8 +16,21 @@ final class PreviewViewImpl with WidgetAdapterImpl implements PreviewView {
   int? get identifier => api.pigeon_instanceManager.getIdentifier(api);
 
   @override
-  Future<void> setController(CameraController controller) =>
-      api.setController(controller.api);
+  Future<void> setController(CameraController? controller) =>
+      api.setController(controller?.api);
+
+  @override
+  Future<void> setImplementationMode(
+    PreviewView$ImplementationMode implementationMode,
+  ) => throw UnimplementedError();
+
+  @override
+  Future<void> setScaleType(PreviewView$ScaleType scaleType) =>
+      api.setScaleType(scaleType.api);
+
+  @override
+  Future<void> setScreenFlashOverlayColor(Color color) =>
+      throw UnimplementedError();
 }
 
 final class PreviewViewChannelImpl extends PreviewViewChannel {
@@ -24,4 +39,8 @@ final class PreviewViewChannelImpl extends PreviewViewChannel {
     final api = PreviewViewProxyApi();
     return PreviewViewImpl.internal(api);
   }
+}
+
+extension PreviewView$ScaleTypeX on PreviewView$ScaleType {
+  PreviewViewScaleTypeApi get api => PreviewViewScaleTypeApi.values[index];
 }